Refactoring StackElement into PartialParse
