+++ ctparse/rule.py
@@ -1,10 +1,22 @@
-import regex
 import logging
-from . types import RegexMatch
+from typing import Callable, Union
 
+import regex
+
+from .types import Artifact, RegexMatch
 
 logger = logging.getLogger(__name__)
 
+
+# A predicate is a callable that returns True if the predicate
+# applies to the artifact
+Predicate = Callable[[Artifact], bool]
+
+# ProductionRule is a function of the kind f(ts: datetime, *args) but this can't be
+# specified with the current mypy syntax
+ProductionRule = Callable[..., Artifact]
+
+rules = {}  # type: Dict[str, Tuple[ProductionRule, Predicate]]
 rules = {}
 
 _regex_cnt = 100  # leave this much space for ids of production types
@@ -32,7 +44,7 @@ _defines = (r'(?(DEFINE)(?<_hour>{regex_hour})(?P<_minute>{regex_minute})'
                 regex_year=_regex_year)
 
 
-def rule(*patterns):
+def rule(*patterns: Union[str, Predicate]):
     def _map(p):
         if type(p) is str:
             # its a regex
@@ -42,9 +54,9 @@ def rule(*patterns):
                 return regex_match(_str_regex[p])
             # test the regex first
             re = r'{defines}(?i)(?P<R{re_key}>{re})'.format(
-                    defines=_defines,
-                    re=p,
-                    re_key=_regex_cnt)
+                defines=_defines,
+                re=p,
+                re_key=_regex_cnt)
             new_rr = regex.compile(
                 # Removed the separator here - leads to more matches,
                 # as now each rule can also match if it is not followed
@@ -75,7 +87,7 @@ def rule(*patterns):
 
     mapped_patterns = [_map(p) for p in patterns]
 
-    def fwrapper(f):
+    def fwrapper(f: ProductionRule) -> ProductionRule:
         def wrapper(ts, *args):
             res = f(ts, *args)
             if res is not None:
@@ -88,22 +100,22 @@ def rule(*patterns):
     return fwrapper
 
 
-def regex_match(r_id):
+def regex_match(r_id) -> Predicate:
     def _regex_match(r):
         return type(r) == RegexMatch and r.id == r_id
     return _regex_match
 
 
-def dimension(dim):
+def dimension(dim) -> Predicate:
     def _dimension(d):
         return isinstance(d, dim)
     return _dimension
 
 
-def predicate(pred):
+def predicate(pred) -> Predicate:
     def _predicate(d):
         return getattr(d, pred, False)
     return _predicate
 
 
-from . time.rules import *  # noqa
+from .time.rules import *  # noqa
